Share the AAC description the cutter and the reverser each read and wrote
Both tools re-encode a track's sound when it cannot be carried across - the
cutter when clips describe their sound differently, the reverser always -
and each does that in its own way. What they had in common was the half
that opens an mp4a sample entry far enough to find the AudioSpecificConfig
an AAC decoder needs, and the half that builds the same nest of MPEG-4
descriptors round the one an encoder hands back: about a hundred and fifty
lines, identical but for the phrase key one refusal threw. That half is
shared/js/aac.js now, importing mp4-boxes for the bytes, and each audio.js
re-exports the two functions so main.js and the tests reach them as before.

The refusal - a description too long for the one-byte length written here,
which no real encoder produces - was `audio.toobig` on the cutter and
`audio.descriptor` on the reverser, the same sentence in fifteen languages
under two names. It is `audio.descriptor` on both; the cutter's fifteen
bodies renamed their span and kept their words.
