+++ locales/ar/tools/trim-video.html
@@ -241,7 +241,7 @@
     <span data-phrase="join.sound">{name} صوته بـ{channels} قناة عند {rate} هرتز والمقطع الأول صوته بـ{firstchannels}
       قناة عند {firstrate} هرتز.</span>
     <span data-phrase="join.soundentry">{name} صوته مُرمَّز على نحو يختلف عن المقطع الأول.</span>
-    <span data-phrase="audio.toobig">خطأ داخلي: وصف الصوت أكبر مما هو متوقع.</span>
+    <span data-phrase="audio.descriptor">خطأ داخلي: وصف الصوت أكبر مما هو متوقع.</span>
     <span data-phrase="audio.noencode">فُكَّ ترميز الصوت لكن لم يعد شيء من المُرمِّز.</span>
     <span data-phrase="nothing.selected">لم يُحدَّد شيء للإبقاء عليه.</span>
     <span data-phrase="encode.toobig">لن يرمّز هذا المتصفح H.264 بحجم {size}. اختر إطاراً أصغر، أو استعمل «أبقِ كل بايت»،
