+++ locales/de/tools/trim-video.html
@@ -248,7 +248,7 @@
     <span data-phrase="join.sound">{name} hat {channels}-kanaligen Ton mit {rate} Hz und der erste Clip hat
       {firstchannels}-kanaligen mit {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} hat anders kodierten Ton als der erste Clip.</span>
-    <span data-phrase="audio.toobig">Interner Fehler: die Tonbeschreibung ist größer als erwartet.</span>
+    <span data-phrase="audio.descriptor">Interner Fehler: die Tonbeschreibung ist größer als erwartet.</span>
     <span data-phrase="audio.noencode">Der Ton wurde dekodiert, aber vom Encoder kam nichts zurück.</span>
     <span data-phrase="nothing.selected">Es ist nichts ausgewählt, was behalten werden soll.</span>
     <span data-phrase="encode.toobig">Dieser Browser wird H.264 nicht in {size} kodieren. Wählen Sie einen kleineren
