+++ locales/es/tools/trim-video.html
@@ -246,7 +246,7 @@
     <span data-phrase="join.sound">{name} tiene sonido de {channels} canales a {rate} Hz y el primer clip lo tiene de
       {firstchannels} canales a {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} tiene el sonido codificado de otra forma que el primer clip.</span>
-    <span data-phrase="audio.toobig">Error interno: la descripción del sonido es más grande de lo esperado.</span>
+    <span data-phrase="audio.descriptor">Error interno: la descripción del sonido es más grande de lo esperado.</span>
     <span data-phrase="audio.noencode">El sonido se descodificó pero del codificador no volvió nada.</span>
     <span data-phrase="nothing.selected">No hay nada seleccionado para quedarse.</span>
     <span data-phrase="encode.toobig">Este navegador no codificará H.264 a {size}. Elige un marco más pequeño, o usa
