+++ locales/fr/tools/trim-video.html
@@ -248,7 +248,7 @@
     <span data-phrase="join.sound">{name} a un son à {channels} canaux en {rate} Hz et le premier clip en a un à
       {firstchannels} canaux en {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} a un son encodé autrement que celui du premier clip.</span>
-    <span data-phrase="audio.toobig">Erreur interne : la description du son est plus grande que prévu.</span>
+    <span data-phrase="audio.descriptor">Erreur interne : la description du son est plus grande que prévu.</span>
     <span data-phrase="audio.noencode">Le son a été décodé mais rien n'est revenu de l'encodeur.</span>
     <span data-phrase="nothing.selected">Il n'y a rien de sélectionné à garder.</span>
     <span data-phrase="encode.toobig">Ce navigateur n'encodera pas du H.264 en {size}. Choisissez un cadre plus petit, ou
