+++ locales/hi/tools/trim-video.html
@@ -245,7 +245,7 @@
     <span data-phrase="join.sound">{name} में {rate} Hz पर {channels} चैनल की आवाज़ है और पहली क्लिप में {firstrate} Hz
       पर {firstchannels} चैनल की।</span>
     <span data-phrase="join.soundentry">{name} की आवाज़ पहली क्लिप से अलग तरीके से एनकोड की गई है।</span>
-    <span data-phrase="audio.toobig">भीतरी गलती: आवाज़ का विवरण उम्मीद से बड़ा है।</span>
+    <span data-phrase="audio.descriptor">भीतरी गलती: आवाज़ का विवरण उम्मीद से बड़ा है।</span>
     <span data-phrase="audio.noencode">आवाज़ डिकोड तो हुई पर एनकोडर से कुछ लौटा ही नहीं।</span>
     <span data-phrase="nothing.selected">रखने के लिए कुछ चुना ही नहीं गया।</span>
     <span data-phrase="encode.toobig">यह ब्राउज़र {size} पर H.264 एनकोड नहीं करेगा। छोटा फ़्रेम चुनिए, या &laquo;हर बाइट
