+++ locales/id/tools/trim-video.html
@@ -252,7 +252,7 @@
     <span data-phrase="join.sound">{name} bersuara {channels} kanal pada {rate} Hz dan klip pertama {firstchannels}
       kanal pada {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} suaranya dikodekan berbeda dari klip pertama.</span>
-    <span data-phrase="audio.toobig">Kesalahan internal: deskripsi suaranya lebih besar daripada yang diperkirakan.</span>
+    <span data-phrase="audio.descriptor">Kesalahan internal: deskripsi suaranya lebih besar daripada yang diperkirakan.</span>
     <span data-phrase="audio.noencode">Suaranya sudah didekode tapi dari pengodenya tidak ada yang kembali.</span>
     <span data-phrase="nothing.selected">Tidak ada yang dipilih untuk disimpan.</span>
     <span data-phrase="encode.toobig">Peramban ini tidak akan mengodekan H.264 pada {size}. Pilih bingkai yang lebih
