+++ locales/it/tools/trim-video.html
@@ -245,7 +245,7 @@
     <span data-phrase="join.sound">{name} ha un audio a {channels} canali a {rate} Hz e il primo clip ce l'ha a
       {firstchannels} canali a {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} ha l'audio codificato diversamente dal primo clip.</span>
-    <span data-phrase="audio.toobig">Errore interno: la descrizione dell'audio è più grande del previsto.</span>
+    <span data-phrase="audio.descriptor">Errore interno: la descrizione dell'audio è più grande del previsto.</span>
     <span data-phrase="audio.noencode">L'audio è stato decodificato ma dal codificatore non è tornato niente.</span>
     <span data-phrase="nothing.selected">Non c'è niente di selezionato da tenere.</span>
     <span data-phrase="encode.toobig">Questo browser non codificherà H.264 a {size}. Scegli un riquadro più piccolo,
