+++ locales/ja/tools/trim-video.html
@@ -229,7 +229,7 @@
     <span data-phrase="join.silent">{name}には音がなく、ほかにはあります。コピーすると途中で無音になってしまいます。</span>
     <span data-phrase="join.sound">{name}の音は{rate} Hzの{channels}チャンネルで、最初のクリップは{firstrate} Hzの{firstchannels}チャンネルです。</span>
     <span data-phrase="join.soundentry">{name}の音は最初のクリップとは違うエンコードです。</span>
-    <span data-phrase="audio.toobig">内部エラー：音声の記述が想定より大きすぎます。</span>
+    <span data-phrase="audio.descriptor">内部エラー：音声の記述が想定より大きすぎます。</span>
     <span data-phrase="audio.noencode">音声はデコードされましたが、エンコーダーから何も返ってきませんでした。</span>
     <span data-phrase="nothing.selected">残すものが何も選ばれていません。</span>
     <span data-phrase="encode.toobig">このブラウザーは{size}のH.264をエンコードしません。もっと小さいフレームを選ぶか、何もエンコードしない「1バイトも残す」を使ってください。</span>
