+++ locales/ko/tools/trim-video.html
@@ -240,7 +240,7 @@
     <span data-phrase="join.silent">{name}에는 소리가 없고 다른 것들에는 있어서, 복사하면 도중에 조용해집니다.</span>
     <span data-phrase="join.sound">{name}은 {rate} Hz의 {channels}채널 소리이고 첫 클립은 {firstrate} Hz의 {firstchannels}채널입니다.</span>
     <span data-phrase="join.soundentry">{name}은 소리가 첫 클립과 다르게 인코딩되어 있습니다.</span>
-    <span data-phrase="audio.toobig">내부 오류: 소리 설명이 예상보다 큽니다.</span>
+    <span data-phrase="audio.descriptor">내부 오류: 소리 설명이 예상보다 큽니다.</span>
     <span data-phrase="audio.noencode">소리는 디코딩되었지만 인코더에서 아무것도 돌아오지 않았습니다.</span>
     <span data-phrase="nothing.selected">남길 것으로 선택된 것이 없습니다.</span>
     <span data-phrase="encode.toobig">이 브라우저는 {size}로 H.264를 인코딩하지 않습니다. 더 작은 프레임을 고르거나, 아무것도 인코딩하지 않는 &laquo;모든 바이트
