+++ locales/nl/tools/trim-video.html
@@ -246,7 +246,7 @@
     <span data-phrase="join.sound">{name} heeft geluid met {channels} kanalen op {rate} Hz en de eerste clip met
       {firstchannels} kanalen op {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} heeft het geluid anders gecodeerd dan de eerste clip.</span>
-    <span data-phrase="audio.toobig">Interne fout: de geluidsbeschrijving is groter dan verwacht.</span>
+    <span data-phrase="audio.descriptor">Interne fout: de geluidsbeschrijving is groter dan verwacht.</span>
     <span data-phrase="audio.noencode">Het geluid is gedecodeerd maar er kwam niets terug van de encoder.</span>
     <span data-phrase="nothing.selected">Er is niets geselecteerd om te houden.</span>
     <span data-phrase="encode.toobig">Deze browser codeert geen H.264 op {size}. Kies een kleiner kader, of neem
