+++ locales/pt/tools/trim-video.html
@@ -244,7 +244,7 @@
     <span data-phrase="join.sound">{name} tem som de {channels} canais a {rate} Hz e o primeiro clipe tem de
       {firstchannels} canais a {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} tem o som codificado de outro jeito que o primeiro clipe.</span>
-    <span data-phrase="audio.toobig">Erro interno: a descrição do som é maior do que o esperado.</span>
+    <span data-phrase="audio.descriptor">Erro interno: a descrição do som é maior do que o esperado.</span>
     <span data-phrase="audio.noencode">O som foi decodificado mas do codificador não voltou nada.</span>
     <span data-phrase="nothing.selected">Não há nada selecionado para guardar.</span>
     <span data-phrase="encode.toobig">Este navegador não vai codificar H.264 em {size}. Escolha um quadro menor, ou use
