+++ locales/tr/tools/trim-video.html
@@ -250,7 +250,7 @@
     <span data-phrase="join.sound">{name} {rate} Hz'de {channels} kanallı ses taşıyor, ilk klip ise {firstrate} Hz'de
       {firstchannels} kanallı.</span>
     <span data-phrase="join.soundentry">{name} sesi ilk klipten farklı kodlanmış.</span>
-    <span data-phrase="audio.toobig">İç hata: ses tanımı beklenenden büyük.</span>
+    <span data-phrase="audio.descriptor">İç hata: ses tanımı beklenenden büyük.</span>
     <span data-phrase="audio.noencode">Ses çözüldü ama kodlayıcıdan geriye bir şey gelmedi.</span>
     <span data-phrase="nothing.selected">Tutulacak hiçbir şey seçilmemiş.</span>
     <span data-phrase="encode.toobig">Bu tarayıcı {size} boyutunda H.264 kodlamayacak. Daha küçük bir çerçeve seçin, ya da
