+++ locales/zh-TW/tools/trim-video.html
@@ -236,7 +236,7 @@
     <span data-phrase="join.silent">{name} 沒有聲音，別的有，拷出來會中途變啞。</span>
     <span data-phrase="join.sound">{name} 的聲音是 {rate} Hz 的 {channels} 聲道，第一段是 {firstrate} Hz 的 {firstchannels} 聲道。</span>
     <span data-phrase="join.soundentry">{name} 的聲音編碼方式跟第一段不一樣。</span>
-    <span data-phrase="audio.toobig">內部錯誤：聲音描述比預期的大。</span>
+    <span data-phrase="audio.descriptor">內部錯誤：聲音描述比預期的大。</span>
     <span data-phrase="audio.noencode">聲音解出來了，但編碼器什麼都沒還回來。</span>
     <span data-phrase="nothing.selected">沒有選中任何要留下的東西。</span>
     <span data-phrase="encode.toobig">這個瀏覽器不會按 {size} 編 H.264。挑個小一點的畫幅，或者用「一個位元組都留著」，那個根本不編碼。</span>
