+++ locales/zh/tools/trim-video.html
@@ -236,7 +236,7 @@
     <span data-phrase="join.silent">{name} 没有声音，别的有，拷出来会中途变哑。</span>
     <span data-phrase="join.sound">{name} 的声音是 {rate} Hz 的 {channels} 声道，第一段是 {firstrate} Hz 的 {firstchannels} 声道。</span>
     <span data-phrase="join.soundentry">{name} 的声音编码方式跟第一段不一样。</span>
-    <span data-phrase="audio.toobig">内部错误：声音描述比预期的大。</span>
+    <span data-phrase="audio.descriptor">内部错误：声音描述比预期的大。</span>
     <span data-phrase="audio.noencode">声音解出来了，但编码器什么都没还回来。</span>
     <span data-phrase="nothing.selected">没有选中任何要留下的东西。</span>
     <span data-phrase="encode.toobig">这个浏览器不会按 {size} 编 H.264。挑个小一点的画幅，或者用「一个字节都留着」，那个根本不编码。</span>
