+++ tools/reverse-video/tool.toml
@@ -32,7 +32,7 @@ css_parts = ["progress"]
 # Shared modules this tool asks for. The same arrangement as css_parts,
 # and for the same reason: a component more than one tool needs and no
 # tool should own. file-picker brings in the drop zone.
-js_parts = ["file-picker", "codec-support", "mp4-reader", "mp4-writer", "video-support", "message-box", "media", "format", "webcodecs", "errors", "mp4-boxes"]
+js_parts = ["file-picker", "codec-support", "mp4-reader", "mp4-writer", "video-support", "message-box", "media", "format", "webcodecs", "errors", "mp4-boxes", "aac"]
 # Where a finished result can be carried without being saved first: the row of
 # links shared/handoff.js shows beside the download once there is one. Slugs of
 # tools whose picker takes what this page produces; checked against the tools
