+++ tools/trim-video/body.html
@@ -242,7 +242,7 @@
     <span data-phrase="join.sound">{name} has {channels}-channel sound at {rate} Hz and the first clip has
       {firstchannels}-channel at {firstrate} Hz.</span>
     <span data-phrase="join.soundentry">{name} has sound encoded differently from the first clip.</span>
-    <span data-phrase="audio.toobig">Internal error: the audio description is larger than expected.</span>
+    <span data-phrase="audio.descriptor">Internal error: the audio description is larger than expected.</span>
     <span data-phrase="audio.noencode">The sound was decoded but nothing came back from the encoder.</span>
     <span data-phrase="nothing.selected">There is nothing selected to keep.</span>
     <span data-phrase="encode.toobig">This browser will not encode H.264 at {size}. Choose a smaller frame, or use
